requirements.txt
allure-python-commons==2.13.5
playwright==1.45.1
pytest==8.2.2
pytest-image-snapshot==0.4.0
python-dotenv ==1.0.1
pillow==10.3.0
requests==2.32.3